Fix iteration numbers and beta values in HLRF
The second output of HLRF, containing the model evaluations in each iteration step, was showing the wrong iteration numbers and the wrong values of the beta.
The number is fixed by capturing the case of HLRF in the evaluation of ObjectiveFunction.
The beta is fixed by using the Standard normal space samples instead of the physical when calling the objective target from HLRF.