Rewrite OptimizationProblem
Working with design variables required to start work on an overhaul of
the Input object. This uses a map to store the members and dependent
properties to parse them into the desired subtypes, e.g. DesignVariables
or RandomVariables. Should in the end be much cleaner than the current
struct approach.