Fix LineData for plotting daptive line sampling
With this LineData can be constructed again as the third output of computeFailureProbability of AdaptiveLineSampling, used only to get nice looking 2d plots. NOTE: the plot function is weird, and takes only the first 2 input quantities for the plot. It is not worth to fix because the code is a mess, but it ws fixed to compare the results of the old stable implementeation with the new clean room implementation in the development branch.